Add and play sfx/music

+++ src/OnyxCs.Gba.AnimEngine/AObject.cs
@@ -4,5 +4,5 @@ namespace OnyxCs.Gba.AnimEngine;
 
 public abstract class AObject
 {
-    public abstract void Execute(AnimationSpriteManager animationSpriteManager, Action<int> soundEventCallback);
+    public abstract void Execute(AnimationSpriteManager animationSpriteManager, Action<ushort> soundEventCallback);
 }
